pyproject.toml
[project]
name = "search-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "babel==2.17.0",
    "beautifulsoup4==4.13.4",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "courlan==1.3.2",
    "dateparser==1.2.1",
    "distro==1.9.0",
    "exceptiongroup==1.3.0",
    "google-api-core==2.25.0",
    "google-api-python-client==2.171.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "googleapis-common-protos==1.70.0",
    "googlesearch-python==1.3.0",
    "h11==0.16.0",
    "htmldate==1.9.3",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "jiter==0.10.0",
    "justext==3.0.2",
    "lxml==5.4.0",
    "lxml-html-clean==0.4.2",
    "mcp==1.9.3",
    "openai==1.85.0",
    "proto-plus==1.26.1",
    "protobuf==6.31.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "regex==2024.11.6",
    "requests==2.32.4",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sse-starlette==2.3.6",
    "starlette==0.47.0",
    "tld==0.13.1",
    "tqdm==4.67.1",
    "trafilatura==2.0.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzlocal==5.3.1",
    "uritemplate==4.2.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.3",
    "validators==0.35.0",
    ]
